Create README.md
Background: 
Our client, a Chief Marketing Officer (CMO) of an online restaurant reservation company, approached us for assistance with two problems. The company operated in over 20 cities and the CMO wanted to know how many sales executives and account managers to assign to each city. Sales executives would convince restaurants to join the portal, while account managers would act as a single point of contact between the company and the restaurants. In addition, the company's performance varied across cities, and the CMO sought advice on developing a city-level marketing strategy.

Problem: 
Our client provided us with two datasets: the month-on-month company performance by city and the market presence of each competitor by month and city. However, the broad requirement could lead to analysis paralysis, which would be unproductive for both parties. Without a clear understanding of which insight is more critical, the project would require significant time and resources.

Solution: 
To avoid analysis paralysis, we focused on one problem at a time. For the first problem, we analyzed the distribution of restaurant partners by city to determine the appropriate number of sales executives and account managers needed. In terms of the second problem, we analyzed the company's market share by city and the average commission per restaurant reservation. We built several visualizations to answer these questions and then used the BCG matrix to divide the cities based on their performance. Our recommendations were as follows:
1.	For cities with high market share and high average commission - Deploy more sales executives to maximize market share.
2.	For cities with low market share and high average commission - Provide discounts on commission and deploy more sales executives while adopting an aggressive marketing strategy.
3.	For cities with high market share and low average commission - Deploy more account managers to establish better relationships with restaurants and help them generate more revenue, resulting in higher commissions for the company.
4.	For cities with low market share and low average commission - Do nothing.
We presented our approach and recommendations to the client, and after an hour-long discussion, the CMO was highly impressed with our output and expressed a desire to rehire us in the future.